P4ADEV-4861-activity-invoking-new-notify-timeline-category-api: moved activity to stream package
